Remove queue exclusivity.
Queues marked without this flag will enable round robin politics
when messages are sent to it. In this way, the system could be use
more daemons which are running at the same time in order to
improve scalability. Daemons who have fallen down won't affect
the whole system as long as there are still more daemons alive.